Split researchmap import responsibilities
